refactor: clean up src classes, simplify notifyBuild, split IntegrationTest
- PipelineLogger: make format private, add debug() gated on PIPELINE_DEBUG env
- BranchPolicy: add @CompileStatic (pure value class, all @NonCPS)
- Remove NotificationPayload and notification-payload.json; notifyBuild now
  builds the message directly from BuildContext
- eventually: replace usage comment with @NamedParams-style Accepts: doc line
  (groovy-all:2.4 on compile classpath predates @NamedParam)
- withParallelMatrix: remove usage comment, keeping code self-explanatory
- Split IntegrationTest into per-domain classes with Java text blocks
