--- resources/com/mkobit/libraryexample/notification-payload.json
@@ -1,1 +1,0 @@
-{"text": "Build {{status}}: {{job}} #{{build}} ({{branch}})"}
